backend/requirements.txt
# ======================
# InferX-ML Backend Dependencies
# ======================

# Core Framework
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Task Queue
celery==5.3.4
redis==5.0.1

# Object Storage
minio==7.2.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1

# Machine Learning - Core
scikit-learn==1.3.2
xgboost==2.0.2
imbalanced-learn==0.11.0

# Machine Learning - Time Series
prophet==1.1.5
statsmodels==0.14.0

# Machine Learning - Deep Learning
tensorflow==2.15.0
# keras is included in tensorflow 2.15.0, no separate install needed

# Machine Learning - Explainability
shap==0.44.0
lime==0.2.0.1

# Image Processing
pillow==10.1.0
opencv-python==4.8.1.78

# API & Validation
marshmallow==3.20.1
python-dotenv==1.0.0
pydantic==2.5.2

# Utilities
tqdm==4.66.1
joblib==1.3.2
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0

# AI/LLM Integration
google-generativeai==0.3.2
groq>=0.4.0

